refactor(torii): move query related files to own folder
