+++ crates/torii/graphql/src/lib.rs
@@ -1,5 +1,6 @@
 mod constants;
 mod object;
+mod query;
 pub mod schema;
 pub mod server;
 pub mod types;
