+++ crates/torii/graphql/src/object/mod.rs
@@ -3,9 +3,7 @@ pub mod component_state;
 pub mod connection;
 pub mod entity;
 pub mod event;
-mod filter;
-pub mod input;
-mod query;
+pub mod inputs;
 pub mod system;
 pub mod system_call;
